+++ andes/models/vsc.py
@@ -51,7 +51,7 @@ class VSCShunt(ACDC2Term):
                          v_str='a',
                          e_str='u * (gsh * v**2 - gsh * v * vsh * cos(a - ash) - '
                                'bsh * v * vsh * sin(a - ash)) - psh',
-                         diag_eps=1e-8,
+                         diag_eps=True,
                          )
         self.vsh = Algeb(info='voltage magnitude behind transformer',
                          tex_name="V_{sh}",
@@ -59,7 +59,7 @@ class VSCShunt(ACDC2Term):
                          v_str='v0',
                          e_str='u * (-bsh * v**2 - gsh * v * vsh * sin(a - ash) + '
                                'bsh * v * vsh * cos(a - ash)) - qsh',
-                         diag_eps=1e-8,
+                         diag_eps=True,
                          )
         self.psh = Algeb(info='active power injection into VSC',
                          tex_name="P_{sh}",
@@ -67,14 +67,14 @@ class VSCShunt(ACDC2Term):
                          v_str='p0 * (mode_s0 + mode_s1)',
                          e_str='u * (mode_s0 + mode_s1) * (p0 - psh) + '
                                'u * (mode_s2 + mode_s3) * (v1 - v2 - vdc0)',
-                         diag_eps=1e-8,
+                         diag_eps=True,
                          )
         self.qsh = Algeb(info='reactive power injection into VSC',
                          tex_name="Q_{sh}",
                          v_str='q0 * (mode_s0 + mode_s2)',
                          e_str='u * (mode_s0 + mode_s2) * (q0 - qsh) + '
                                'u * (mode_s1 + mode_s3) * (v0 - v)',
-                         diag_eps=1e-8,
+                         diag_eps=True,
                          )
         self.pdc = Algeb(info='DC power injection',
                          tex_name="P_{dc}",